updating

+++ models/Product.js
@@ -20,7 +20,7 @@ Product.init(
       allowNull: false,
     },
     price: {
-      type: Datatypes.INTEGER,
+      type: Datatypes.DECIMAL,
       allowNull: false,
       validate: {
         isDecimal: true,
